+++ slice.go
@@ -274,7 +274,7 @@ func KeyBy[K comparable, V any](collection []V, iteratee func(V) K) map[K]V {
 
 // Associate returns a map containing key-value pairs provided by transform function applied to elements of the given slice.
 // If any of two pairs would have the same key the last one gets added to the map.
-// The returned map preserves the entry iteration order of the original array.
+// The order of keys in returned map is not specified and is not guaranteed to be the same from the original array.
 func Associate[T any, K comparable, V any](collection []T, transform func(T) (K, V)) map[K]V {
 	result := make(map[K]V)
 
@@ -284,6 +284,14 @@ func Associate[T any, K comparable, V any](collection []T, transform func(T) (K,
 	}
 
 	return result
+}
+
+// Associate returns a map containing key-value pairs provided by transform function applied to elements of the given slice.
+// If any of two pairs would have the same key the last one gets added to the map.
+// The order of keys in returned map is not specified and is not guaranteed to be the same from the original array.
+// Alias of Associate().
+func SliceToMap[T any, K comparable, V any](collection []T, transform func(T) (K, V)) map[K]V {
+	return Associate(collection, transform)
 }
 
 // Drop drops n elements from the beginning of a slice or array.